Improved scaffold - 21/03/2024

+++ pokemon_base.py
@@ -150,7 +150,7 @@ class Pokemon(ABC): # pylint: disable=too-few-public-methods, too-many-instance-
         """
         return self.battle_power
 
-    def attack(self, other_pokemon) -> int:
+    def attack(self, other_pokemon) -> float:
         """
         Calculates and returns the damage that this Pokemon inflicts on the
         other Pokemon during an attack.
